model_pipeline/backend/requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-community
langchain-openai
langchain-core
ragas
datasets
sentence-transformers
python-dotenv
langgraph
pandas
numpy
pyyaml
pinecone[grpc]
google-cloud-secret-manager>=2.16.0
google-cloud-storage
openpyxl
structlog
langfuse
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
httpx>=0.24.0
pytest-mock>=3.11.0